Bring in Olivier Hardy's 1.4 changes:
New version (released January 2013). What's new in SPAGeDi ver 1.4?

Compared to the previous version 1.3, there are several major
improvements in this version:

1.  SPAGeDi 1.4 can compute a new estimator of pairwise relatedness
between individuals using AFLP or RAPD markers when phenotypes (presence
or absence of a band) are affected by non genetic factors. To this end,
the broad-sense heritability (H2) of each marker must first be estimated
using replicated individuals. These H2 estimates are provided at the end
of the data file and used to compute corrected estimates of pairwise
relatedness coefficients. Details about the statistics used can be found
in Ley & Hardy (2013) who showed that H2 is an appropriate measure of
the reliability of AFLP markers and that H2 can be much lower than one
(the value expected for a "perfect" dominant marker) for the majority of
AFLP markers.

2.  SPAGeDi 1.4 can compute for each population and locus the
effective number of alleles (Nielsen et al. 2003), the rarefied allelic
richness for which the subsample size can be specified, and the
inbreeding coefficient with significance tests.

3.  SPAGeDi 1.4 can test differentiation statistics for each pair of
populations.